requirements.txt
affine==2.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.61.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
rasterio==1.4.4
six==1.17.0
Werkzeug==3.1.5
